+++ lib/location_map_picker.dart
@@ -1,0 +1,4 @@
+library pin_google_map;
+
+export 'src/location_map_picker.dart';
+