+++ lib/pin_google_map.dart
@@ -1,4 +1,0 @@
-library pin_google_map;
-
-export 'src/pin_google_map.dart';
-